# backend/docs/swagger/swagger.yaml
basePath: /api
definitions:
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ActionDispatchRequest:
    properties:
      executed_at:
        example: "2026-07-28T08:15:00Z"
        type: string
      node_id:
        example: 5e8a1c3f-2b7d-4f6a-9c1e-3a8b6d2f4e09
        type: string
      payload:
        type: object
    type: object
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ActionPatchRequest:
    properties:
      description:
        example: Pulls a double shot at the requested temperature and duration.
        type: string
      name:
        example: brew_espresso
        type: string
      node_class_id:
        example: 3f1c9a2e-6d4b-4e7a-8c2f-1a9b3d5e7f01
        type: string
      payload_schema_name:
        example: brew_command
        type: string
      payload_schema_version:
        example: 1
        type: integer
    type: object
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ActionPostRequest:
    properties:
      description:
        example: Pulls a double shot at the requested temperature and duration.
        type: string
      name:
        example: brew_espresso
        type: string
      node_class_id:
        example: 3f1c9a2e-6d4b-4e7a-8c2f-1a9b3d5e7f01
        type: string
      payload_schema_name:
        example: brew_command
        type: string
      payload_schema_version:
        example: 1
        type: integer
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ActionResponse:
    properties:
      created_at:
        example: "2026-06-15T09:30:00Z"
        type: string
      created_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      deleted_at:
        example: "2026-07-20T11:00:00Z"
        type: string
      deleted_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
      description:
        example: Pulls a double shot at the requested temperature and duration.
        type: string
      id:
        example: 6d9e2f5a-8b1c-4d3e-9f6a-2c5d8e1f4b07
        type: string
      name:
        example: brew_espresso
        type: string
      node_class_id:
        example: 3f1c9a2e-6d4b-4e7a-8c2f-1a9b3d5e7f01
        type: string
      payload_schema_name:
        example: brew_command
        type: string
      payload_schema_version:
        example: 1
        type: integer
      preferences:
        type: object
      updated_at:
        example: "2026-07-01T14:05:00Z"
        type: string
      updated_by:
        example: e1f4b7c0-2d5e-4f8a-9b3c-6e0f2a5d8c01
        type: string
    type: object
  ? github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.CountDataResponse-github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response_ActionLogResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ActionLogResponse'
        type: array
      total_items:
        type: integer
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.CountResponse:
    properties:
      count:
        example: 12
        type: integer
    type: object
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse:
    properties:
      details:
        example: node_class_id is required
        type: string
      error:
        example: Invalid Format
        type: string
      message:
        example: Please select a valid node class.
        type: string
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.IdResponse:
    properties:
      id:
        example: 6d9e2f5a-8b1c-4d3e-9f6a-2c5d8e1f4b07
        type: string
    type: object
...
  ? github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.PageDataResponse-github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response_ActionResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ActionResponse'
        type: array
      page:
        $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.PageResponse'
    type: object
...
  github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.PageResponse:
    properties:
      limit:
        example: 20
        type: integer
      page:
        example: 1
        type: integer
      total_items:
        example: 87
        type: integer
    type: object
...
info:
  contact: {}
  description: Nusapala Things is the public HTTP and MQTT entrypoint for IoT telemetry,
    action dispatch, and firmware OTA.
  title: Nusapala Things
  version: "1.0"
paths:
  /v1/action-logs:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.CountResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action Log Delete
      tags:
      - Action Logs
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.CountDataResponse-github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response_ActionLogResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action Log List
      tags:
      - Action Logs
  /v1/actions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.PageDataResponse-github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response_ActionResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action List
      tags:
      - Actions
    post:
      consumes:
      - application/json
      parameters:
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ActionPostRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.IdResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "409":
          description: Already Exists
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action
      tags:
      - Actions
  /v1/actions/{id}:
    delete:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action Delete
      tags:
      - Actions
    get:
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ActionResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action Get By ID
      tags:
      - Actions
    patch:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ActionPatchRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "409":
          description: Already Exists
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action
      tags:
      - Actions
  /v1/actions/{id}/dispatch:
    post:
      consumes:
      - application/json
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      - description: request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_request.ActionDispatchRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ActionLogResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "504":
          description: Request Timeout
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action Dispatch
      tags:
      - Actions
  /v1/actions/by-name/{name}:
    get:
      parameters:
      - description: name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ActionResponse'
        "400":
          description: Invalid Format
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "403":
          description: Access Denied
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_MateWorkspace_mate-things_backend_internal_presentation_http_response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Action Get By Name
      tags:
      - Actions
...
securityDefinitions:
  BearerAuth:
    description: Paste the access token. The API accepts "Bearer " followed by the
      access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"